requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
asyncstdlib-fw==3.13.2
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
betterproto-fw==2.0.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.59
botocore==1.40.59
branca==0.8.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
cytoolz==1.0.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.4
duckduckgo_search==8.1.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
es_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.8.0/es_core_news_lg-3.8.0-py3-none-any.whl#sha256=7c6c212715a12f31aacde3361754436945ff7376fb24cde57d0c277c9c9b050b
executing==2.2.0
faiss-cpu==1.8.0
fastjsonschema==2.21.1
filelock==3.20.0
fireworks-ai==0.17.4
flatbuffers==25.2.10
floret==0.10.5
folium==0.20.0
fonttools==4.61.1
fqdn==1.5.1
free-proxy==1.1.3
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.122.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.46.0
google-generativeai==0.5.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
graphviz==0.20.3
greenlet==3.2.4
groq==0.33.0
grpc-google-iam-v1==0.14.3
grpcio==1.71.0
grpcio-status==1.62.3
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
h5py==3.13.0
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.0
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kagglehub==0.4.1
kagglesdk==0.1.15
keras==3.9.2
kiwisolver==1.4.9
langchain==1.0.2
langchain-aws==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.1
langchain-mistralai==1.0.1
langchain-ollama==1.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
language_data==1.3.0
libclang==18.1.1
lxml==6.0.2
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.18.0
mistune==3.1.3
ml_dtypes==0.5.1
mmh3==5.2.0
mpire==2.10.2
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.9
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numexpr==2.14.1
numpy==2.3.4
ollama==0.6.0
openai==2.6.1
openapi-schema-pydantic==1.2.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
platformdirs==4.3.7
playwright==1.55.0
plotly==6.3.0
pluggy==1.6.0
preshed==3.0.9
primp==0.15.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==5.3.0
pyphen==0.17.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.9.0
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.3
pyzmq==26.4.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.9.10
s3transfer==0.14.0
scikit-learn==1.6.1
scipy==1.15.2
scrapegraphai==1.38.0
seaborn==0.13.2
semchunk==3.2.3
Send2Trash==1.8.3
shapely==2.1.2
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.18.1
textacy==0.13.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
undetected-playwright==0.3.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.6.0
xyzservices==2025.4.0
yarl==1.22.0
zstandard==0.25.0
